requirements.txt
numpy 
pandas
matplotlib
seaborn
python-dotenv
mysql-connector-python
pymysql
scikit-learn
# -e .